requirements.txt
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Assets==2.0
Flask-Bootstrap==3.3.7.1
itsdangerous==1.1.0
Jinja2==2.11.3
libsass==0.20.1
MarkupSafe==1.1.1
numpy==1.20.1
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
Pillow==8.1.0
pyScss==1.3.7
python-dotenv==0.15.0
six==1.15.0
visitor==0.1.3
webassets==2.0
Werkzeug==1.0.1
